refactor: use double for quantity management

+++ core/src/main/java/com/ttdat/core/application/commands/product/UpdateProductTotalQuantityCommand.java
@@ -15,5 +15,5 @@ public class UpdateProductTotalQuantityCommand {
     @TargetAggregateIdentifier
     String productId;
 
-    Integer quantity;
+    Double quantity;
 }
