+++ product-service/src/main/java/com/ttdat/productservice/api/dto/common/ProductDTO.java
@@ -23,7 +23,7 @@ public class ProductDTO {
 
     String description;
 
-    Integer totalQuantity;
+    Double totalQuantity;
 
     String imageUrl;
